🌱 fix: split remediation and drilldown modal components
- RemediationConsole.tsx (792 -> 152 lines): extracted shared types to
  RemediationConsole.types.ts, run/shell lifecycle state to the
  useRemediationRun hook, and presentational sections to
  RemediationConsole.parts.tsx. No behavior change.
- DrillDownModal.tsx (463 -> 154 lines): extracted the lazy view
  registry/router to DrillDownModal.views.tsx and modal chrome
  (error boundary, loading fallback, icon/status helpers, breadcrumbs,
  footer hints) to DrillDownModal.parts.tsx. The drill-down stack
  navigation contract via useDrillDown() is unchanged.

Fixes #21788
Part of #21791 (DrillDownModal slice only; SecretDrillDown and
ComplianceDrillDown are tracked as follow-up child issues to keep this
PR reviewable).
